src/counterpart/texture.py
"""
Photometric texture baking module using PyTorch3D.

References:
- PyTorch3D: https://pytorch3d.org/
- FLAME model: https://flame.is.tue.mpg.de/
"""
from typing import List, Dict, Tuple
import numpy as np
import torch
from PIL import Image

class TextureBaker:
    """Stub for texture baking using PyTorch3D."""
    def __init__(self, device: str = "cuda"):
        # TODO: Initialize PyTorch3D renderer
        self.device = device
        # self.renderer = ...

    def bake_texture(self, 
                    images: List[str],
                    flame_params: Dict[str, np.ndarray],
                    uv_coords: np.ndarray) -> np.ndarray:
        """Bake photometric texture from multiple images.
        Args:
            images: List of image paths
            flame_params: Dictionary containing FLAME parameters:
                - 'shape': Shape coefficients (shape (300,))
                - 'exp': Expression coefficients (shape (50,))
                - 'pose': Pose parameters (shape (6,))
            uv_coords: UV coordinates for texture mapping (shape (N, 2))
        Returns:
            Baked texture as numpy array (shape (H, W, 3))
        """
        # Load all images
        loaded_images = [load_image(img) for img in images]
        
        # TODO: Implement actual texture baking
        # For now, return a dummy texture
        texture_size = 1024
        return np.zeros((texture_size, texture_size, 3), dtype=np.uint8)

def load_image(image_path: str) -> np.ndarray:
    """Load and preprocess image for texture baking.
    Args:
        image_path: Path to input image
    Returns:
        Preprocessed image as numpy array (shape (H, W, 3))
    """
    # TODO: Implement actual image loading and preprocessing
    # For now, return a dummy image
    return np.zeros((256, 256, 3), dtype=np.uint8)

def save_texture(texture: np.ndarray, output_path: str):
    """Save baked texture to file.
    Args:
        texture: Baked texture as numpy array (shape (H, W, 3))
        output_path: Path to save texture
    """
    # TODO: Implement actual texture saving
    # For now, just save as PNG
    Image.fromarray(texture).save(output_path) 